requirements.txt
mcp>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
email-validator>=2.0.0
